fix: update manage account with buttons for delete and transfer
