+++ frontend/src/Components/inputs/DropdownControl.tsx
@@ -2,10 +2,6 @@ import { Dispatch, SetStateAction } from 'react';
 
 interface DropdownControlProps {
     label?: string;
-    /**
-     * Optional attribute is used for enabling the lable for being the first item displayed in the list
-     */
-    useLabel?: boolean;
     setState?: Dispatch<SetStateAction<string>>;
     customCallback?: (value: string) => void;
     enumType: Record<string, string>;
@@ -16,8 +12,7 @@ export default function DropdownControl({
     label,
     setState: callback,
     customCallback,
-    enumType,
-    useLabel = false
+    enumType
 }: DropdownControlProps) {
     return (
         <label className="form-control">
@@ -33,14 +28,10 @@ export default function DropdownControl({
                     }
                 }}
             >
-                {label && !useLabel ? (
+                {label && (
                     <option value="" disabled>
                         {label}
                     </option>
-                ) : label && useLabel ? (
-                    <option value={label}>{label}</option>
-                ) : (
-                    ''
                 )}
                 {Object.entries(enumType).map(([key, value]) => (
                     <option key={key} value={value}>
